chore: split and move `utils` to `cargo_near_build` (#200)
- **chore: move cargo_near::util::print ->
cargo_near_build::pretty_print**
- **chore: move `util::{copy, force_cononicalize_dir}` to
`cargo_near_build::fs`**
- **chore: move `wasm32_exists` logic to
`cargo_near_build::cargo_native::target`**
- **chore: move `dylib_extention` to `cargo_near_build::cargo_native`**
- **chore: move `util::{...near types...}` ->
`cargo_near_build::types::near::{...}`**
- **chore: parameterize `CompilationArtifact` with `T: ArtifactType`**
- **chore: move `util::{... build funcs ...}` ->
`cargo_near_build::cargo_native::compile::run`**
- **chore: move `util::extract_abi_entries` ->
`cargo_near_build::near::abi`**
