+++ cargo-near-build/Cargo.toml
@@ -7,3 +7,13 @@ version = "0.1.0"
 eyre = "0.6.12"
 camino = "1.1.1"
 colored = "2.0"
+dunce = "1"
+log = "0.4"
+bs58 = "0.5"
+hex = "0.4.3"
+sha2 = "0.10"
+cargo_metadata = "0.18.1"
+symbolic-debuginfo = "8.8"
+libloading = "0.8.3"
+serde_json = "1.0"
+near-abi = { version = "0.4.0", features = ["__chunked-entries"] }
