+++ cargo-near-build/src/near/mod.rs
@@ -1,0 +1,1 @@
+pub mod abi;
