+++ cargo-near-build/src/types/cargo/mod.rs
@@ -1,1 +1,2 @@
+// TODO: consider making mod non-pub
 pub mod manifest_path;
