+++ cargo-near-build/src/types/mod.rs
@@ -1,2 +1,3 @@
 pub mod cargo;
 pub mod color_preference;
+pub mod near;
