+++ cargo-near/Cargo.toml
@@ -24,8 +24,6 @@ eula = false
 
 [dependencies]
 cargo-near-build = { version = "0.1.0", path = "../cargo-near-build" }
-bs58 = "0.5"
-hex = "0.4.3"
 camino = "1.1.1"
 cargo_metadata = "0.18.1"
 clap = { version = "4.0.18", features = ["derive", "env"] }
@@ -35,11 +33,8 @@ log = "0.4"
 rustc_version = "0.4"
 serde = "1.0.197"
 serde_json = "1.0"
-sha2 = "0.10"
-symbolic-debuginfo = "8.8"
 schemars = "0.8"
 near-abi = { version = "0.4.0", features = ["__chunked-entries"] }
-libloading = "0.8.3"
 zstd = "0.13"
 color-eyre = "0.6"
 inquire = "0.7"
